fix: 걸음수측정 무한로딩 버그 해결
