+++ src/navigation/StackNavigator.js
@@ -1,4 +1,4 @@
-import React, {useEffect, useState} from 'react';
+import React from 'react';
 import TabNavigator from './TabNavigator';
 import {
   MypageMainScreen,
@@ -25,63 +25,10 @@ import {
 } from '../screens';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import LoginScreen from '../screens/Login/LoginScreen';
-import {usePedometer} from '../feature/pedometer/recoil/usePedometer';
-import {
-  accelerometer,
-  SensorTypes,
-  setUpdateIntervalForType,
-} from 'react-native-sensors';
-import {saveData, STEP_COUNT} from '../data/LocalStorage';
-import AsyncStorage from '@react-native-async-storage/async-storage';
 
 const Stack = createNativeStackNavigator();
-setUpdateIntervalForType(SensorTypes.accelerometer, 400);
 
 export default function StackNavigator() {
-  const [xAcceleration, setXAcceleration] = useState(0);
-  const [yAcceleration, setYAcceleration] = useState(0);
-  const [zAcceleration, setZAcceleration] = useState(0);
-  const [magnitudePrevious, setMagnitudePrevious] = useState(0);
-
-  const {stepCount, setStepCount} = usePedometer();
-
-  async function initStepCount() {
-    const value = await AsyncStorage.getItem(STEP_COUNT);
-    if (value == null) value = 0;
-    setStepCount(parseInt(value));
-  }
-
-  useEffect(() => {
-    initStepCount();
-
-    const subscription = accelerometer
-      .pipe(data => data)
-      .subscribe(speed => {
-        setXAcceleration(speed.x);
-        setYAcceleration(speed.y);
-        setZAcceleration(speed.z);
-      });
-    return () => {
-      subscription.unsubscribe();
-    };
-  }, []);
-
-  useEffect(() => {
-    const magnitude = Math.sqrt(
-      Math.pow(xAcceleration, 2) +
-        Math.pow(yAcceleration, 2) +
-        Math.pow(zAcceleration, 2),
-    );
-
-    const magnitudeDelta = magnitude - magnitudePrevious;
-    setMagnitudePrevious(() => magnitude);
-
-    if (magnitudeDelta > 2) {
-      setStepCount(prevSteps => prevSteps + 1);
-      saveData(STEP_COUNT, stepCount.toString());
-    }
-  }, [xAcceleration, yAcceleration, zAcceleration]);
-
   return (
     <Stack.Navigator
       initialRouteName="OnBoarding"
